chore: nitpicks (#6)
- use *const bool across FFI boundary instead of *const u8
- suppress hundreds of compiler warnings (but move unused & untested
  functions into an "untested" module)
- Clean up copy pasta mess in some comments
