--- barretenberg/bb_rs/src/barretenberg_api/traits.rs
@@ -1,8 +1,0 @@
-pub trait SerializeBuffer {
-    fn to_buffer(&self) -> Vec<u8>;
-}
-
-pub trait DeserializeBuffer {
-    type Slice;
-    fn from_buffer(val: Self::Slice) -> Self;
-}
