+++ barretenberg/bb_rs/src/barretenberg_api/untested/bn254.rs
@@ -1,8 +1,6 @@
-use super::{
-    bindgen,
-    models::Fr,
-    traits::{DeserializeBuffer, SerializeBuffer},
-};
+use crate::barretenberg_api::bindgen;
+use crate::barretenberg_api::utils::{DeserializeBuffer, SerializeBuffer};
+use crate::models::Fr;
 
 /// Compute the square root of a field element in BN254's Fr field
 /// Returns Some(sqrt) if the square root exists, None otherwise
@@ -12,10 +10,10 @@ pub unsafe fn bn254_fr_sqrt(input: &Fr) -> Option<Fr> {
         input.to_buffer().as_slice().as_ptr(),
         result_buf.as_mut_ptr(),
     );
-    
+
     // First byte indicates whether a square root exists
     let is_sqrt = result_buf[0] == 1;
-    
+
     if is_sqrt {
         // Extract the square root from the remaining 32 bytes
         let mut sqrt_buf = [0u8; 32];
@@ -24,4 +22,4 @@ pub unsafe fn bn254_fr_sqrt(input: &Fr) -> Option<Fr> {
     } else {
         None
     }
-} 
+}
