+++ barretenberg/bb_rs/src/barretenberg_api/untested/mod.rs
@@ -1,0 +1,10 @@
+pub mod aes;
+pub mod blake2s;
+pub mod bn254;
+pub mod ecdsa;
+pub mod grumpkin;
+pub mod pedersen;
+pub mod poseidon2;
+pub mod schnorr;
+pub mod secp256k1;
+pub mod srs;
