+++ barretenberg/bb_rs/src/barretenberg_api/untested/pedersen.rs
@@ -1,9 +1,6 @@
-use super::{
-    bindgen,
-    models::{Fr, Point},
-    traits::SerializeBuffer,
-};
-use crate::barretenberg_api::traits::DeserializeBuffer;
+use crate::barretenberg_api::bindgen;
+use crate::barretenberg_api::utils::{DeserializeBuffer, SerializeBuffer};
+use crate::models::{Fr, Point};
 
 pub unsafe fn pedersen_commit(inputs: &[Fr], hash_index: u32) -> Point {
     let mut output: <Point as DeserializeBuffer>::Slice = [0; 64];
@@ -33,26 +30,26 @@ pub unsafe fn pedersen_hashes(inputs: &[Vec<Fr>], hash_index: u32) -> Vec<Fr> {
             flattened_inputs.push(*fr);
         }
     }
-    
+
     // The C++ function processes pairs, so we expect half as many results as flattened inputs
     let expected_results = inputs.len();
     // Allocate buffer for vector serialization: 4 bytes (count) + expected_results * 32 bytes (Fr data)
     let mut output_buffer = vec![0u8; 4 + expected_results * 32];
-    
+
     bindgen::pedersen_hashes(
         flattened_inputs.to_buffer().as_slice().as_ptr(),
         hash_index.to_be_bytes().as_ptr() as *const u32,
         output_buffer.as_mut_ptr(),
     );
-    
+
     // Parse the output buffer: first 4 bytes are the count, then Fr values
     let count = u32::from_be_bytes([
         output_buffer[0],
-        output_buffer[1], 
+        output_buffer[1],
         output_buffer[2],
-        output_buffer[3]
+        output_buffer[3],
     ]) as usize;
-    
+
     let mut results = Vec::new();
     for i in 0..count {
         let start = 4 + i * 32; // Skip the 4-byte count prefix
@@ -61,13 +58,13 @@ pub unsafe fn pedersen_hashes(inputs: &[Vec<Fr>], hash_index: u32) -> Vec<Fr> {
         fr_data.copy_from_slice(&output_buffer[start..end]);
         results.push(Fr::from_buffer(fr_data));
     }
-    
+
     results
 }
 
 pub unsafe fn pedersen_hash_buffer(inputs: &[u8], hash_index: u32) -> Fr {
     let mut output: <Fr as DeserializeBuffer>::Slice = [0; 32];
-    
+
     // Handle empty buffer case - C++ might not handle zero-length buffers correctly
     if inputs.is_empty() {
         // For empty buffer, we can return a deterministic hash based on just the hash_index
@@ -80,10 +77,10 @@ pub unsafe fn pedersen_hash_buffer(inputs: &[u8], hash_index: u32) -> Fr {
         );
         return Fr::from_buffer(output);
     }
-    
+
     // Serialize the buffer with length prefix as expected by the C++ function
     let serialized_buffer = inputs.to_buffer();
-    
+
     bindgen::pedersen_hash_buffer(
         serialized_buffer.as_ptr(),
         hash_index.to_be_bytes().as_ptr() as *const u32,
