+++ barretenberg/bb_rs/src/barretenberg_api/untested/srs.rs
@@ -1,0 +1,5 @@
+use crate::barretenberg_api::bindgen;
+
+pub unsafe fn init_srs(points_buf: &[u8], num_points: u32) {
+    bindgen::srs_init_srs(points_buf.as_ptr(), &num_points.to_be());
+}
